==> Analysis/03-compare-patch-protocols.Rmd <==
---
title: "03-compare-patch-protocols"
author: "Ariel Marcy"
date: "2018-10-15"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = '../eco-rodents')
```

# Compare big and small patch protocols
Our study asked a methodological question: does it make a difference if we use big areas of patch semi-landmarks versus breaking it up with curve semi-landmarks? Or in other words, will a big patch of landmarks pick up different biological patterns than an area broken up by curve landmarks defined along sutures?

We hypothesized that the big patch protocol would pick up convergent evolution while the small patch + suture curve protocol would better detect phylogentically-conserved shapes. 

The heatmaps provide a statistically rigorous way to visualize the relative variation of landmarks across a protocol. In other words, the colors show us which parts of the skull vary most and if the protocols emphasize different patterns.

### Load packages, functions, and data from previous steps
```{r message = FALSE}
library(png)  # needed to work with screenshots
library(colorspace)  # plot colors
# Load libraries (and install if necessary) for landvR package
if(!require(devtools)) install.packages("devtools")
if(!require(dispRity)) install.packages("dispRity")
if(!require(landvR)) install_github("TGuillerme/landvR")
library(data.table)
library(vegan)
library(geomorph)
source("../Functions/utilities.R")  # custom functions
load(file = "../Data/Processed/02-main-data-both-protocols.rda")
```

## Landmark variation heatmaps for big patch protocol
Here we use Dr Thomas Guillerme's new package, `landvR` to see how landmarks vary within and between the two protocols. Much of the below was adapted from [this vignette written by Dr Guillerme](https://cdn.rawgit.com/TGuillerme/landvR/8a6a6bd5/inst/vignettes/Landmark_partition_test.html).
```{r}
# Find mean configuration - i.e. mean shape in dataset
big.consensus <- (select.procrustes(big.shape, selector = mean))[[1]]

# Measure spherical coordinates differences from the mean for each specimen for each landmark
diff.from.mean <- coordinates.difference(coordinates = big.shape, reference = big.consensus, type = "spherical")
```

To test the hypothesis that each landmark's variation from the mean configuration is above expected, we can use the PC axes from a PCA.
```{r}
# Ordinate the data
twoD.shape <- two.d.array(big.shape)
ordination <- stats::prcomp(twoD.shape)

# Force the symmetric component of shape into class "gpagen" (required for variation.range)
big.gpagen.shape <- list()
big.gpagen.shape$coords <- big.shape
big.gpagen.shape$consensus <- big.consensus
class(big.gpagen.shape) <- "gpagen"

# Measure extremes of variation from mean on PC1
big.PC1.var <- variation.range(big.gpagen.shape, return.ID = FALSE, axis = 1, ordination = ordination, type = "spherical")
```

The next step is to make a heatmap showing the size of variation and direction of movement for each landmark in the dataset, comparing between the min and max specimens along PC1.
```{r}
# Wrap specimens on the tangent space
big.wrap.PCA <- plotTangentSpace(big.shape, verbose = FALSE)

# Select extreme specimens (the ones that make the warp-meshes) on PC1
big.hypothetical.1 <- big.wrap.PCA$pc.shapes[[1]]  # PC1 min
big.hypothetical.2 <- big.wrap.PCA$pc.shapes[[2]]  # PC1 max
```

**Repeat the steps immediately above for PC2**
The pc.shapes[1] and [2] correspond to PC1min and PC1max. Therefore we can adjust the same big.wrap.PCA to plot the heatmaps for PC2. 
```{r}
# Measure extremes of variation from mean on PC2
big.PC2.var <- variation.range(big.gpagen.shape, return.ID = FALSE, axis = 2, ordination = ordination, type = "vector")  # can also try "spherical"

# Select extreme specimens on PC2
big.hypothetical.3 <- big.wrap.PCA$pc.shapes[[3]]  # PC2 min
big.hypothetical.4 <- big.wrap.PCA$pc.shapes[[4]]  # PC2 max
```

## Repeat landmark variation heatmap for small patch protocol 
```{r}
# Find mean configuration - i.e. mean shape in dataset
sm.consensus <- (select.procrustes(sm.shape, selector = mean))[[1]]

# Measure spherical coordinates differences from the mean for each specimen for each landmark
diff.from.mean <- coordinates.difference(coordinates = sm.shape, reference = sm.consensus, type = "spherical")

# Ordinate the data
twoD.shape <- two.d.array(sm.shape)
ordination <- stats::prcomp(twoD.shape)

# Force the symmetric component of shape into class "gpagen" (required for variation.range)
sm.gpagen.shape <- list()
sm.gpagen.shape$coords <- sm.shape
sm.gpagen.shape$consensus <- sm.consensus
class(sm.gpagen.shape) <- "gpagen"

## PC1
# Measure extremes of variation from mean on PC1
sm.PC1.var <- variation.range(sm.gpagen.shape, return.ID = FALSE, axis = 1, ordination = ordination, type = "spherical")

# Wrap specimens on the tangent space
sm.wrap.PCA <- plotTangentSpace(sm.shape, verbose = FALSE)

# Select extreme specimens (the ones that make the warp-meshes)
sm.hypothetical.1 <- sm.wrap.PCA$pc.shapes[[1]]
sm.hypothetical.2 <- sm.wrap.PCA$pc.shapes[[2]]

## PC2
# Measure extremes of variation from mean on PC2
sm.PC2.var <- variation.range(big.gpagen.shape, return.ID = FALSE, axis = 2, ordination = ordination, type = "vector")  # can also try "spherical"

# Select extreme specimens on PC2
sm.hypothetical.3 <- sm.wrap.PCA$pc.shapes[[3]]  # PC2 min
sm.hypothetical.4 <- sm.wrap.PCA$pc.shapes[[4]]  # PC2 max
```

## Make heatmap plot comparisons
```{r}
## Big patch protocol PC1
# Dorsal view
open3d(windowRect = c(0, 0, 600, 600))
big.PC1.var.plot <- procrustes.var.plot(big.hypothetical.1, big.hypothetical.2, col = heat.colors, col.val = big.PC1.var[, "radius"], labels = F)
view3d(phi = 0)  # flip cranium
rgl.snapshot("../Data/Processed/heatmap_PC1_big_dor.png")  # save screenshot for multipanel plot

# Lateral view
open3d(windowRect = c(0, 0, 600, 600))
big.PC1.var.plot <- procrustes.var.plot(big.hypothetical.1, big.hypothetical.2, col = heat.colors, col.val = big.PC1.var[, "radius"], labels = F)
view3d(phi = 90)
rgl.snapshot("../Data/Processed/heatmap_PC1_big_lat.png")

## Small patch protocol PC1
# Dorsal view
open3d(windowRect = c(0, 0, 600, 600))
sm.PC1.var.plot <- procrustes.var.plot(sm.hypothetical.1, sm.hypothetical.2, col = heat.colors, col.val = sm.PC1.var[, "radius"], labels = F)
view3d(phi = 0)  # flip cranium
rgl.snapshot("../Data/Processed/heatmap_PC1_sm_dor.png")  # save screenshot for multipanel plot

# Lateral view
open3d(windowRect = c(0, 0, 600, 600))
sm.PC1.var.plot <- procrustes.var.plot(sm.hypothetical.1, sm.hypothetical.2, col = heat.colors, col.val = sm.PC1.var[, "radius"], labels = F)
view3d(phi = 90)
rgl.snapshot("../Data/Processed/heatmap_PC1_sm_lat.png")

## Big patch protocol PC2
# Dorsal view
open3d(windowRect = c(0, 0, 600, 600))
big.PC2.var.plot <- procrustes.var.plot(big.hypothetical.3, big.hypothetical.4, col = heat.colors, col.val = big.PC2.var[, "length"], labels = F)
view3d(phi = 0)  # flip cranium
rgl.snapshot("../Data/Processed/heatmap_PC2_big_dor.png")

# Lateral view
open3d(windowRect = c(0, 0, 600, 600))
big.PC2.var.plot <- procrustes.var.plot(big.hypothetical.3, big.hypothetical.4, col = heat.colors, col.val = big.PC2.var[, "length"], labels = F)
view3d(phi = 90)
rgl.snapshot("../Data/Processed/heatmap_PC2_big_lat.png")

## Small patch protocol PC2
# Dorsal view
open3d(windowRect = c(0, 0, 600, 600))
sm.PC2.var.plot <- procrustes.var.plot(sm.hypothetical.3, sm.hypothetical.4, col = heat.colors, col.val = sm.PC2.var[, "length"], labels = F)
view3d(phi = 0)  # flip cranium
rgl.snapshot("../Data/Processed/heatmap_PC2_sm_dor.png")

# Lateral view
open3d(windowRect = c(0, 0, 600, 600))
sm.PC2.var.plot <- procrustes.var.plot(sm.hypothetical.3, sm.hypothetical.4, col = heat.colors, col.val = sm.PC2.var[, "length"], labels = F)
view3d(phi = 90)
rgl.snapshot("../Data/Processed/heatmap_PC2_sm_lat.png")
```

### Prep images for multi-panel figure TK
This step assumes ImageMagik is installed on your computer
```{r}
system("mogrify ../Data/Processed/heatmap_*.png -trim ../Data/Processed/*.png")  # removes whitespace (requires ImageMagik)
PC1_big_dor <- readPNG("../Data/Processed/heatmap_PC1_big_dor.png")
PC1_big_lat <- readPNG("../Data/Processed/heatmap_PC1_big_lat.png")
PC1_sm_dor <- readPNG("../Data/Processed/heatmap_PC1_sm_dor.png")
PC1_sm_lat <- readPNG("../Data/Processed/heatmap_PC1_sm_lat.png")
PC2_big_dor <- readPNG("../Data/Processed/heatmap_PC2_big_dor.png")
PC2_big_lat <- readPNG("../Data/Processed/heatmap_PC2_big_lat.png")
PC2_sm_dor <- readPNG("../Data/Processed/heatmap_PC2_sm_dor.png")
PC2_sm_lat <- readPNG("../Data/Processed/heatmap_PC2_sm_lat.png")
```

### Plot multi-panel heatmap figure
ERROR: memory issues when trying to add legend within setEPS. Works when lines 107-177 are run through the console, plot is "zoomed" from RStudio, and then screenshot from that window
```{r}
#setEPS()  # sets up plot export
#postscript("../Data/Results/Figure1_Heatmaps.eps")

#mat <- matrix(c(1,2,3,4,5,6,7,8), 4, 2, byrow = TRUE)  # 4 rows, 2 cols
mat <- matrix(c(1,2,3,4), 2, 2, byrow = TRUE)  # 2 rows, 2 cols
layout(mat, widths = c(1, 1, 1, 1), heights = rep(1, 4)) 

## PC1
# 1) Small patch PC1 dorsal
par(mar = c(0, 0, 1, 0))
plot(c(0, dim(PC1_sm_dor)[2]), c(0, dim(PC1_sm_dor)[1]), type = "n", axes = F, xlab = "", ylab = "", asp = T) 
rasterImage(PC1_sm_dor, 0, 0, dim(PC1_sm_dor)[2], dim(PC1_sm_dor)[1])
text(0, dim(PC1_sm_dor)[1]-10, "A", cex = 2)

# 2) Big patch PC1 dorsal
par(mar = c(0, 0, 1, 0))
plot(c(0, dim(PC1_big_dor)[2]), c(0, dim(PC1_big_dor)[1]), type = "n", axes = F, xlab = "", ylab = "", asp = T) 
rasterImage(PC1_big_dor, 0, 0, dim(PC1_big_dor)[2], dim(PC1_big_dor)[1])
text(0, dim(PC1_big_dor)[1]-10, "E", cex = 2)

# 3) Small patch PC1 lateral
par(mar = c(0, 0, 1, 0))
plot(c(0, dim(PC1_sm_lat)[2]), c(0, dim(PC1_sm_lat)[1]), type = "n", axes = F, xlab = "", ylab = "", asp = T) 
rasterImage(PC1_sm_lat, 0, 0, dim(PC1_sm_lat)[2], dim(PC1_sm_lat)[1])
text(0, dim(PC1_sm_lat)[1]-10, "B", cex = 2)

# 4) Big patch PC1 lateral
par(mar = c(0, 0, 0, 0))
plot(c(0, dim(PC1_big_lat)[2]), c(0, dim(PC1_big_lat)[1]), type = "n", axes = F, xlab = "", ylab = "", asp = T) 
rasterImage(PC1_big_lat, 0, 0, dim(PC1_big_lat)[2], dim(PC1_big_lat)[1])
text(0, dim(PC1_big_lat)[1]-10, "F", cex = 2)

## PC2
# 5) Small patch PC2 dorsal
par(mar = c(0, 0, 0, 0))
plot(c(0, dim(PC2_sm_dor)[2]), c(0, dim(PC2_sm_dor)[1]), type = "n", axes = F, xlab = "", ylab = "", asp = T) 
rasterImage(PC2_sm_dor, 0, 0, dim(PC2_sm_dor)[2], dim(PC2_sm_dor)[1])
text(0, dim(PC2_sm_dor)[1]-10, "C", cex = 2)

# 6) Big patch PC2 dorsal
par(mar = c(0, 0, 0, 0))
plot(c(0, dim(PC2_big_dor)[2]), c(0, dim(PC2_big_dor)[1]), type = "n", axes = F, xlab = "", ylab = "", asp = T) 
rasterImage(PC2_big_dor, 0, 0, dim(PC2_big_dor)[2], dim(PC2_big_dor)[1])
text(0, dim(PC2_big_dor)[1]-10, "G", cex = 2)

# 7) Small patch PC2 lateral
par(mar = c(0, 0, 0, 0))
plot(c(0, dim(PC2_sm_lat)[2]), c(0, dim(PC2_sm_lat)[1]), type = "n", axes = F, xlab = "", ylab = "", asp = T) 
rasterImage(PC2_sm_lat, 0, 0, dim(PC2_sm_lat)[2], dim(PC2_sm_lat)[1])
text(0, dim(PC2_sm_lat)[1]-10, "D", cex = 2)

# 8) Big patch PC2 lateral
par(mar = c(0, 0, 0, 0))
plot(c(0, dim(PC2_big_lat)[2]), c(0, dim(PC2_big_lat)[1]), type = "n", axes = F, xlab = "", ylab = "", asp = T) 
rasterImage(PC2_big_lat, 0, 0, dim(PC2_big_lat)[2], dim(PC2_big_lat)[1])
text(0, dim(PC2_big_lat)[1]-10, "H", cex = 2)
```

## Plot PCAs for each protocol
PCAs allow visual comparison of how differently the patch protocols sort the shape data. To assist, we wrote a function, `PlotByGroup()`, to match colors or other plotting attributes to groups within the dataset. These don't impact the analysis; they are just visualization aids.

Here, we'll use this function to look at the morphospace colored by genus. Since Mastacomys is technically within the genus Pseudomys (Smissen & Rowe 2018), we'll make an "Effective Genus" column.
```{r}
# Make an Effective Genus column
big.info$EGenus <- big.info$Genus
big.info$EGenus[which(big.info$Genus == "Mas")] <- "Pse"
sm.info$EGenus <- sm.info$Genus
sm.info$EGenus[which(sm.info$Genus == "Mas")] <- "Pse"
```

### Set up colors
Colors for subsequent figures are set up according to column "EGenus". Colors roughtly correspond to phylogenetics such that similar color hues indicate closer evolutionary relationships.
```{r}
# Make gradients for each clade
Uro <- sequential_hcl(5, "Greens 3")  # Uromyini clade
Con <- sequential_hcl(5, "Purples 3")  # Conilurus clade
Pse <- sequential_hcl(5, "Blues 3")  # Pseudomys clade
Hyd <- sequential_hcl(5, "Reds 3")  # Hydromini clade
Pog <- sequential_hcl(5, "Heat 2")  # Pogonomys
Inv <- sequential_hcl(5, "Light Grays")  # Mus and Rattus

# Colors in order of phylogeny
genera.phylo <- c("Mus", "Rat", "Pog", "Xer", "Hyd", "Not", "Pse", "Zyz", "Leg", "Uro", "Mel", "Lep", "Mes", "Con")

col.phylo <- c(Inv[3], Inv[1], Pog[3], Hyd[1], Hyd[2], Pse[1], Pse[2], Pse[4], Pse[3], Uro[1], Uro[2], Con[3], Con[2], Con[1])

names(col.phylo) <- genera.phylo

# Colors in order of alphabetical genus names
col.gen <- col.phylo[match(sort(unique(big.info$EGenus)), names(col.phylo))]
sp.col.gen <- col.phylo[match(as.factor(big.info$EGenus), names(col.phylo))]  # color vector for all data
```

### PCA plot of big patch protocol
PCAs are very human-friendly way of visualizing shape variation in 2 dimensions. Usually, the proportion of variance for each PC axis is reported in percent.
```{r}
# Generate big patch protocol PCA
big.pca <- plotTangentSpace(big.shape, groups = sp.col.gen, axis1 = 1, axis2 = 2, verbose = T)

# Get proportion of variance for PC1 and PC2 in format for plots
big.PCs <- big.pca$pc.summary$importance
big.PC1 <- round(big.PCs[2, 1] * 100, digits = 1)  # % with 1 decimal
big.PC2 <- round(big.PCs[2, 2] * 100, digits = 1)

# Generate small patch protocol PCA
sm.pca <- plotTangentSpace(sm.shape, groups = sp.col.gen, axis1 = 1, axis2 = 2, verbose = T, label = sm.info$CatNum)

# Get proportion of variance for PC1 and PC2 in format for plots
sm.PCs <- sm.pca$pc.summary$importance
sm.PC1 <- round(sm.PCs[2, 1] * 100, digits = 1)
sm.PC2 <- round(sm.PCs[2, 2] * 100, digits = 1)
```

### Multi-panel PCA plot of both protocols
```{r}
setEPS()  # sets up plot export
postscript("../Data/Results/Figure2_PCA_protocol_compare.eps")
mat <- matrix(c(1,1,2,2), 2, 2, byrow = TRUE)  # 2 rows, 2 columns
layout(mat, widths = c(1, 1), heights = c(1, 1)) 

# 1) PCA for small patch protocol
par(mar = c(4, 4, 1, 1))
plot(x = sm.pca$pc.scores[, 1],
     y = sm.pca$pc.scores[, 2],
     xlab = paste("PC1 (", sm.PC1, "%)", sep = ""), 
     ylab = paste("PC2 (", sm.PC2, "%)", sep = ""),
     xlim = c(-0.11, 0.163),
     ylim = c(-0.09, 0.05),
     asp = TRUE,
     col = sp.col.gen, 
     bg = sp.col.gen,
     pch = 16, 
     cex = 1.5, 
     cex.axis = 1.3, 
     cex.lab = 1.3)
legend("topright", legend = genera.phylo, col = col.phylo, pch = 16, cex = 1)
text(-0.11, 0.045, "Small patch", cex = 1.5)

# 2) PCA for big patch protocol
par(mar=c(4, 4, 1, 1))  # sets the margins
plot(x = big.pca$pc.scores[, 1],
     y = big.pca$pc.scores[, 2],
     xlab = paste("PC1 (", big.PC1, "%)", sep = ""), 
     ylab = paste("PC2 (", big.PC2, "%)", sep = ""),
     xlim = c(-0.11, 0.163),
     ylim = c(-0.09, 0.05),
     asp = TRUE, 
     col = sp.col.gen, 
     bg = sp.col.gen,
     pch = 16, 
     cex = 1.5, 
     cex.axis = 1.3, 
     cex.lab = 1.3)
legend("topright", legend = genera.phylo, col = col.phylo, pch = 16, cex = 1)
text(-0.11, 0.045, "Big patch", cex = 1.5)

dev.off()
```

## Mantel test
A way to test if the two protocols provide a statistically significant different presentations of specimens.
```{r}
mantel(dist(sm.pca$pc.scores), dist(big.pca$pc.scores))
```
The Mantel test finds a highly significant (p < 0.001) relationship between the two PC score distance matrices, with statistic = 0.9989. Meaning they are very similar distance matrices, as expected. 

## Compare performance of patch protocols in ANOVA by genus and centroid size
The patch protocols look very similar on the PCAs, which suggests that they do not capture shape variation differently from one another. To test this, we performed a Procrustes ANOVA on shape ~ centroid size * genus.

### Run shape ~ genus * Csize ANOVAs on each protocol:  
Big patch protocol:
```{r}
big.gdf <- geomorph.data.frame(shape = big.shape, genus = big.info$EGenus, csize = big.info$Csize)
big.procD <- procD.lm(shape ~ csize * genus, data = big.gdf)
summary(big.procD)
```

Small patch protocol:
```{r}
sm.gdf <- geomorph.data.frame(shape = sm.shape, genus = sm.info$EGenus, csize = sm.info$Csize)
sm.procD <- procD.lm(shape ~ csize * genus, data = sm.gdf)
summary(sm.procD)
```
These ANOVAs are also very similar. Therefore, we decided to use only the small patch protocol as this protocol captured more information (sutures versus patch points) and had less asymmetry than the big patch protocol. It also resembles the more common method of landmarking.

### Save intermediate data
Only small patch data from now on.
```{r}
# Remove "sm." from shape and metadata matrix names for increased readability
shape <- sm.shape
info <- sm.info
PCA <- sm.pca
PCs <- sm.PCs
pt.names <- sm.pt.names

save(shape, info, PCA, PCs, col.gen, sp.col.gen, col.phylo, genera.phylo, pt.names, file = "../Data/Processed/03-sm-PCA-with-color.rda")
```
