==> Analysis/03-prepare-data.Rmd <==
---
title: "03-prepare-data"
author: "Ariel Marcy"
date: "12/11/2019"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE)
knitr::opts_knit$set(root.dir = '../eco-rodents')
```

## Prepare metadata, graphing vectors, and the phylogenetic tree
These data will be needed for all subsequent analyses, tables, and figures. The rda file created at the end will be needed for all future scripts. 

### Load packages, functions, and data
```{r message = FALSE}
library(geomorph)   # CRAN v4.0.6
library(stringr)    # CRAN v1.5.0
library(data.table) # CRAN v1.14.8
library(colorspace) # CRAN v2.1-0  
library(phylotate)  # CRAN v1.3  
library(geiger)     # CRAN v2.0.11  

# This sets the wd to local
library(rstudioapi) # Safely Access the RStudio API
setwd(dirname(getActiveDocumentContext()$path))

source("../Functions/utilities.R")  # custom functions
load(file = "../Data/Processed/02-no-error-data-sm.rda")  # small patch shape data w/o error
```

## Attach all needed information to the metadata
```{r}
# Remove unnecessary columns
drops <- c("Patch", "Rep", "All")
info <- sm.info[, !(names(sm.info) %in% drops)]

# Add trait info from Native Mice and Rats (Breed & Ford 2007)
traits <- read.csv("../Data/Processed/in_ex_traits.csv", header = TRUE)
traits$Taxa <- paste(str_sub(traits$Genus, 1, 1), str_sub(traits$Species, 1, 3), sep = "_")  # make matching Taxa column
info <- merge(info, traits, by = "Taxa", sort = TRUE)  # merge trait datasets

info <- info[order(info$Order), ]  # original order / same as shape

# Rename Genus & Species columns so they have their original short column name, long form Genus and Species columns keep Genus.y and Species.y
names(info)[names(info) == "Genus.x"] <- "Genus"
names(info)[names(info) == "Species.x"] <- "Species"

# Add long form names ("Genus_species")
info$FullName <- paste(info$Genus.y, info$Species.y, sep = "_")

# Add info relating to old vs new endemics (vs invasives)
info$Endemic <- "Old"
info$Endemic[which(info$Clade == "Rattini")] <- "New"
info$Endemic[which(info$Clade == "invRattini")] <- "Inv"
info$Endemic[which(info$Clade ==  "Mus")] <- "Inv"


```

## Prepare shape dataset dimnames for phylogenetic analyses
Rename shape dataset dimnames to only "Genus_species". This is so we can re-order the shape data to match the tree later on. (We can still reference the CatNum in the metadata.)
```{r}
shape <- sm.shape
#make sure this is the same as modules so the re-name works for the modules as well
dimnames(shape)[[3]] == dimnames(module_shape$molar$shape)[[3]]


dimnames(shape)[[3]] <- as.list(paste(info$FullName))

for (i in length(module_list)){
  
  dimnames(module_shape[[i]]$shape)[[3]] <- as.list(paste(info$FullName))
  names(module_shape[[i]]$csize) <- as.list(paste(info$FullName))
  
  
}

#Check check
dimnames(shape)[[3]] == dimnames(module_shape[[i]]$shape)[[3]]


```

### Handle *Mastacomys*
Since *Mastacomys* is technically within the genus *Pseudomys* (Smissen & Rowe 2018), we'll make an "Effective Genus" or "EGenus" column in the metadata where this genus is classified under *Pseudomys* for analyses that use genera.
```{r}
info$EGenus <- info$Genus  # make Effective Genus column
info$EGenus[which(info$Genus == "Mas")] <- "Pse"  # Mas becomes Pseudomys
```

## Create colors based on phylogenetic clades in Aplin (2014)
Colors for subsequent figures are set up according to column "EGenus". Colors correspond to phylogenetics - i.e. when different ancestors arrived in Australia. Each genus gets its own color. 

We make three different vectors of color:
```{r}
# Make gradients for the genera in each clade
Uro <- sequential_hcl(5, "Greens 3")  # Uromys clade
Mes <- sequential_hcl(5, "Purples 3")  # Mesembriomys clade
Pse <- sequential_hcl(5, "Blues 3")  # Pseudomys clade
Hyd <- sequential_hcl(5, "Reds 3")  # Hydromys clade
Pog <- sequential_hcl(5, "Heat 2")  # Pogonomys clade
Inv <- sequential_hcl(5, "Light Grays")  # Mus and Rattus clades

# 1) Genus color by phylogeny
genera.phylo <- c("Mus", "Rat", "Pog", "Xer", "Hyd", "Not", "Pse", "Zyz", "Leg", "Uro", "Mel", "Lep", "Mes", "Con")
col.phylo <- c(Inv[3], Inv[1], Pog[3], Hyd[1], Hyd[2], Pse[1], Pse[2], Pse[4], Pse[3], Uro[1], Uro[2], Mes[3], Mes[2], Mes[1])
names(col.phylo) <- genera.phylo 

# 2) Color by alphabetical genus name
col.gen <- col.phylo[match(sort(unique(info$EGenus)), names(col.phylo))] 

# 3) Color vector for all data (specimen level)
sp.col.gen <- col.phylo[match(as.factor(info$EGenus), names(col.phylo))]
sp.col.gen[which(info$Species %in% c("nor", "rat"))] <- "black"  # account for invasive Rattus
```

### Create point and color vectors for species legends in phylo order
```{r}
# Put color and point legends in phylogenetic order
phylo.names <- c("Mus_mus", "Rat_rat", "Rat_nor", "Rat_lut", "Rat_leu", "Rat_tun", "Rat_vil", "Rat_sor", "Rat_fus", "Pog_mol", "Xer_myo", "Hyd_chr", "Not_cer", "Not_fus", "Not_mit", "Not_ale", "Pse_pat", "Pse_gra", "Pse_ora", "Pse_sho", "Pse_des", "Pse_apo", "Pse_aus", "Pse_hig", "Pse_her", "Pse_nov", "Pse_del", "Pse_fus", "Zyz_arg", "Zyz_ped", "Leg_for", "Uro_cau", "Mel_cer", "Mel_cap", "Mel_bur", "Lep_con", "Mes_gou", "Con_pen")

# Points for legend
pch.unique.spp <- PointOutDiffSpp(phylo.names)  # pch by unique spp in a genus
pch.unique.spp[which(phylo.names == "Pse_fus")] <- 2  # handle Mastacomys
names(pch.unique.spp) <- phylo.names

# Points coded for each individual specimen in dataset
info$Name <- paste(info$EGenus, info$Species, sep = "_")
phylo.order <- match(sort(unique(info$Name)), phylo.names)
pch.spp <- PlotByGroup(info, "Name", pch.unique.spp[phylo.order])  

# Make color vector for legends by species
col.taxa <- match(str_sub(phylo.names, 1, 3), names(col.gen))
col.unique.spp <- col.gen[col.taxa]
col.unique.spp[which(phylo.names %in% c("Rat_nor", "Rat_rat"))] <- "black"  # account for invasive Rattus
names(col.unique.spp) <- phylo.names

# Give short names back to phylo.names
phylo.names <- paste(str_sub(phylo.names, 1, 1), str_sub(phylo.names, 5, 7))
phylo.names[which(phylo.names == "P fus")] <- "M fus"  # handle Mastacomys
```

# Prepare data for phylogenetic analyses
First, the shape data have to be averaged by species. This step has to be done before plotting evolutionary allometry and working with the phylogeny. Second, the phylogeny has to be trimmed to the species in the shape data as the commands in `geomorph` require the tree and datasets to have exactly the same set of species.

## 1) Calculate mean shape and centroid size for each species
The mean shape function, `mshape()` outputs a 2D array. Therefore, the `for` loop below stores mean shapes in a 2D array format recognized by `geomorph`: (n x [p x k]), where n is number of specimens, p is number of landmark points, and k is number of dimensions. Finally, we convert to a 3D array with format (p x k x n) since a 3D array is required by most `geomorph` functions. 
```{r}
# For loop to subset shape data by species and calculate mean shape
mean.shapes.2D <- NULL  # initiate blank data frame
full.names <- unique(info$FullName)  # list of species

for (i in 1:length(full.names)) {  
        is.taxa <- which(info$FullName == full.names[i])
        shape.for.mean.2D <- shape[, , is.taxa]
        mean.shapes.2D <- rbind(mean.shapes.2D, mshape(shape.for.mean.2D))  # output cols must = 3 for 3D
}

mean.shapes <- arrayspecs(mean.shapes.2D, 325, 3)  # convert to 3D array with 325 landmarks and 3 for 3D landmarks in this study



#Now same thing but a nested loop for the modules
mean_module_shapes_2D <- NULL
Mean_module_shapes <- module_list

for (i in 1:length(module_list)){
  
  for (j in 1:length(full.names)) {  
    
        is.taxa <- which(info$FullName == full.names[j])
        shape.for.mean.2D <- module_shape[[i]]$shape[, , is.taxa]
        mean_module_shapes_2D <- rbind(mean_module_shapes_2D, mshape(shape.for.mean.2D))  # output cols must = 3 for 3D
      }
  
  #Here I'm soft coding the numbers of landmarks (hard-coded in Ariel's code above)
  Mean_module_shapes[[i]] <- arrayspecs(mean_module_shapes_2D, length(module_shape[[i]]$shape[,,1][,1]), 3)  # convert to 3D array with relevant 3D landmarks in the module
  #need to reset this, otherwise it adds up in the j loop
  mean_module_shapes_2D <- NULL
  }

```

#some checks - the r-pls values of the separately gpa'ed modules and the equivalent partition in the whole-shape gpa should be very high

```{r}

# lm.modules <- read.csv("../Data/Processed/Landmark_modules.csv", sep = ",", header = TRUE)
# mean_full_gpa_partitions <- list( basicran = mean.shapes[which(lm.modules[, 4] == "basicran"), ,], molar = mean.shapes[which(lm.modules[, 4] == "molar"), ,], orbital = mean.shapes[which(lm.modules[, 4] == "orbital"), ,], rostrum = mean.shapes[which(lm.modules[, 4] == "rostrum"), ,], vault = mean.shapes[which(lm.modules[, 4] == "vault"), ,])
# 
# names(mean_full_gpa_partitions) == names(Mean_module_shapes)
# 
# for (i in 1: length(names(Mean_module_shapes)) ){
#   
#   
#  print( integration.test(mean_full_gpa_partitions[[i]], Mean_module_shapes[[i]])$r.pls)
# 
# }


```

#also check that the configurations make shape sense
```{r}

open3d()
mfrow3d(2,3, sharedMouse = TRUE)

cols=c("red", "blue", "purple", "orange", "pink")

for (i in 1: length(names(Mean_module_shapes)) ){
  
spheres3d(Mean_module_shapes[[i]][,,2], col= cols[i], radius=0.01 )
texts3d(0,0,0.1, texts = names(Mean_module_shapes)[i])
  
  next3d()
  
}
  
```

Create metadata and put mean shape and metadata in alphabetical order.
```{r}
# Alphabetical order for mean shape
dimnames(mean.shapes)[[3]] <- full.names  # Genus_species format
mean.shapes <- mean.shapes[, , sort(dimnames(mean.shapes)[[3]])]  # ABC order


#For modules

for (i in 1: length(names(Mean_module_shapes)) ){
  
dimnames(Mean_module_shapes[[i]])[[3]] <- full.names
Mean_module_shapes[[i]] <- Mean_module_shapes[[i]][, , sort(dimnames(Mean_module_shapes[[i]])[[3]])]  # ABC order

}

# Alphabetical order for mean shape metadata
info.means <- info[!duplicated(info$Taxa), ]  # make metadata
drops <- c("CatNum", "Csize", "Order")  # remove outdated columns
info.means <- info.means[, !(names(info.means) %in% drops)]
info.means <- info.means[order(info.means$FullName), ]  # ABC order
```

### Calculate mean centroid size
We need to update the centoid size column for info.means to contain the *mean* centroid size by taxa. Centroid size for all 317 specimens is stored in the info.phylo metadata table.
```{r}
# Calculate mean centroid size from original Csize data and store in info.means
cs.means <- aggregate(info[, which(colnames(info) == "Csize")], list(Taxa = info$FullName), mean)  # aggregate also alphabetizes it
info.means$MeanCsize <- cs.means[, 2]  # store CS in dataframe

#for modules. This aggregates the csizes that stayed associated with each module above (in the module_shape object). Note that the aggregate function spits out a taxon name column and the actual size values, hance [,2] above and $x below

mean_modules_shapesize <- module_list

for (i in 1: length(names(Mean_module_shapes)) ){
  
mean_modules_shapesize[[i]]$shape <- Mean_module_shapes[[i]]
mean_modules_shapesize[[i]]$csize <- aggregate(module_shape[[i]]$csize, list(Taxa=info$FullName), mean)[,2]
names(mean_modules_shapesize[[i]]$csize) <- aggregate(module_shape[[i]]$csize, list(Taxa=info$FullName), mean)[,1]
}

#check check
names(mean_modules_shapesize$molar$csize) == dimnames(mean_modules_shapesize$basicran$shape)[[3]]

#Checksto make sure the csizes all match up with the full data, just a correlation.

for (i in 1: length(names(Mean_module_shapes)) ){
  

print(cor(mean_modules_shapesize[[i]]$csize, info.means$MeanCsize))

}

#also - are the largest species really the  largest and the smallest the smallest? Yes :-)

names(which(mean_modules_shapesize$vault$csize==max(mean_modules_shapesize$vault$csize))) == info.means$FullName[ which(info.means$MeanCsize==max(info.means$MeanCsize))]

names(which(mean_modules_shapesize$basicran$csize==min(mean_modules_shapesize$basicran$csize)))==info.means$FullName[ which(info.means$MeanCsize==min(info.means$MeanCsize))]

```

### Names, colors, and points in alphabetical order
Also need to fix tricky Mastacomys (should be Pseudomys)
```{r}
# Correct for order where Mastacomys is in the dataset
names(col.unique.spp)[which(names(col.unique.spp) == "Pse_fus")] <- "Mas_fus"

alpha.col <- col.unique.spp[order(names(col.unique.spp))]  # colors
alpha.pch <- pch.unique.spp[order(names(col.unique.spp))]  # points
alpha.names <- dimnames(mean.shapes)[[3]]  # names

#these alpha names are also useable for the modules:
dimnames(mean.shapes)[[3]] == dimnames(mean_modules_shapesize$basicran$shape)[[3]]

```

## 2) Prepare phylogenetic tree
The most recent tree for Australian rodents is from Smissen & Rowe 2018. The tree is a fossil calibrated ultrametric tree from BEAST2 for Hydromyini, which includes all the Australo-Papuan old endemics in Murinae. This tree adds 6 more species from Australia to the Smissen & Rowe (2018) phylogeny. 
```{r}
raw.tree <- read_annotated("../Data/Processed/Marcy-BEAST01.con.tre")
```

### Extract only Genus and Species names from the tree
The tree file has uneven amounts of information for each entry, some entries have catalog numbers and others do not. Therefore we needed a `for` loop to grab the first two elements from a list, which always corresponded to Genus and Species.

We reassign only Genus and Species names (in the form "Genus_species") to the tree so that it is easier to compare to the shape data. 
```{r}
no.cat.tree <- raw.tree  # initiate new, modified tree with no CatNums
label.list <- strsplit(raw.tree$tip.label, "_")  # separate info into list

# For loop finds Genus & Species and makes tip name "Genus_species"
for (i in 1:length(label.list)) {
        label <- unlist(label.list[i])
        no.cat.tree$tip.label[i] <- paste(label[1], label[2], sep = "_")
}
```

### Re-label node
Since the 2018 Smissen & Rowe tree focused primarily on New Guinea species, it does not have all of the Australian species in our shape dataset. However, it does include sister species with the same relative branch length to other genera in the study. Therefore, we decided to re-label one of the nodes to the sister Australian species. 

**Relabeled node:**
_Pogonomys loriae_ (NG) to _Pogonomys mollipilosus_ (Aus)
```{r}
relabel.tree <- no.cat.tree
P.lor <- which(str_detect(no.cat.tree$tip.label, "Pogonomys_loriae"))
relabel.tree$tip.label[P.lor] <- paste("Pogonomys_mollipilosus")
```

**Correct Rattus**
R. lutreolus and R. leucopus were switched relative to Rowe et al 2011 and other recent phylogenies. Likely due to a simple labelling error in phylogenetic preparation. Here we correct it. 
```{r}
R.lut <- which(str_detect(relabel.tree$tip.label, "Rattus_lutreolus"))
R.leu <- which(str_detect(relabel.tree$tip.label, "Rattus_leucopus"))

relabel.tree$tip.label[R.lut] <- paste("Rattus_leucopus")
relabel.tree$tip.label[R.leu] <- paste("Rattus_lutreolus")
```

### Prune the tree and shape datasets to match each other
Now that the names in the Smissen & Rowe tree can be compared to our info.phylo metadata, we can prune both datasets to have the same species.
```{r}
# First, prune the tree of species not in the shape data
concord <- name.check(relabel.tree, relabel.tree$tip.label, dimnames(mean.shapes)[[3]])  # check concordance
aus.tree <- drop.tip(relabel.tree, concord$tree_not_data)  # prune tree to Aus sample
```

Next, subset and re-order the metadata and shape datasets to match the order of species found in the tree. 
```{r}
# Subset datasets
to.remove <- name.check(aus.tree, aus.tree$tip.label, dimnames(mean.shapes)[[3]])

#this namecheck should be identical to the species in the module dataset adn the "to remove" object should work for both - and it is
name.check(aus.tree, aus.tree$tip.label, dimnames(mean_modules_shapesize$molar$shape)[[3]])$data_not_tree == name.check(aus.tree, aus.tree$tip.label, dimnames(mean.shapes)[[3]])$data_not_tree

info.means.tree <- info.means[!is.element(info.means$FullName, to.remove$data_not_tree), ]  # metadata
mean.shapes.tree <- mean.shapes[, , !is.element(info.means$FullName, to.remove$data_not_tree)]  # shape data

#for the modules:
module_sizeshape_tree <- module_list

for (i in 1:length(module_list)){

module_sizeshape_tree[[i]]$shape <- mean_modules_shapesize[[i]]$shape[, , !is.element(info.means$FullName, to.remove$data_not_tree)]  # shape data    
module_sizeshape_tree[[i]]$csize <- mean_modules_shapesize[[i]]$csize[!is.element(info.means$FullName, to.remove$data_not_tree)]  # csize

}

# Re-order datasets to match tree tip order
tree.order <- match(aus.tree$tip.label, info.means.tree$FullName)
info.means <- info.means.tree[tree.order, ]  # metadata
mean.shapes <- mean.shapes.tree[, , tree.order]  # shape data

#for the modules. Again, the csize is just a vector here 

for (i in 1:length(module_list)){

module_sizeshape_tree[[i]]$shape <- mean_modules_shapesize[[i]]$shape[, , tree.order] # shape data    
module_sizeshape_tree[[i]]$csize <- mean_modules_shapesize[[i]]$csize[tree.order] #csize 

}

#again, this should all check out in terms of largest and smallest specimen. To they are all sorted in the same way.

names(which(module_sizeshape_tree$vault$csize==max(module_sizeshape_tree$vault$csize))) == info.means$FullName[ which(info.means$MeanCsize==max(info.means$MeanCsize))]

names(which(module_sizeshape_tree$molar$csize==min(module_sizeshape_tree$molar$csize)))==info.means$FullName[ which(info.means$MeanCsize==min(info.means$MeanCsize))]

#Match with phylogeny still
names(module_sizeshape_tree$orbital$csize) == aus.tree$tip.label
  
#Names and csizes are also still correctly associated between module and full dataset

for (i in 1: length(names(Mean_module_shapes)) ){
  
print(cor(module_sizeshape_tree[[i]]$csize, info.means$MeanCsize))

}

#There have been a lot of intermediate datasets I don't want saved, so I am here re-naming the final processed dataset into something to take into analyses
mean_modules <- module_sizeshape_tree

```

### Save intermediate data
```{r}
# Metadata and shape coordinates for all specimens
save(info, shape, file = "../Data/Processed/03-main-data.rda")

# Graphing vectors
save(genera.phylo, col.phylo, col.gen, col.unique.spp, sp.col.gen, phylo.names, pch.spp, col.unique.spp, pch.unique.spp, alpha.names, alpha.col, alpha.pch, file = "../Data/Processed/03-graphing-vectors.rda")

# Data for phylogenetic analyses, including the tree as well as metadata and shape coordinates for mean specimens
save(info.means, mean.shapes, mean_modules, aus.tree, file = "../Data/Processed/03-tree-data.rda")
```
