==> Analysis/04-compare-PCAs.Rmd <==
---
title: "04-compare_PCAs"
author: "Ariel Marcy & Vera Weisbecker"
date: "8/15/2019"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = '../eco-rodents')
```

# Plot PCAs for allometric and shape residual datasets
Plot PCAs for allometric shape and shape residual, provide coloration and pch symbols by locomotion and diet, and visualize changes in shape residual PCA when *Notomys* species are removed. These plots are relevant to the heatplots created in script 05 / shown in Figure 2. 

Creates multipanel Figure 1.

Compares morphospace similarity of the two datasets with Mantel tests and a correlation test.

Creates supplementary Figure 1, a screeplot of the the PCs for each dataset.

Produces shape residual data used in later analyses, stored in rda file 04.

### Load packages, functions, and data
Tree, mean shape, and metadata for phylogenetic analyses come from the allometry rodents script 04. 
```{r message = FALSE}
library(geomorph)   # CRAN v4.0.6
library(vegan)      # CRAN v2.6-4  
library(dplyr)      # CRAN v1.1.3
library(colorspace) # CRAN v2.1-0

# This sets the wd to local
library(rstudioapi) # Safely Access the RStudio API
setwd(dirname(getActiveDocumentContext()$path))

source("../Functions/utilities.R")  # custom functions
load(file = "../Data/Processed/03-main-data.rda")
load(file = "../Data/Processed/03-tree-data.rda")
load(file = "../Data/Processed/03-graphing-vectors.rda")
```

# 1) Export Figure 1
Create complex, multipanel plot of PCAs from different datasets

## A) PCA of mean shapes in full shape dataset (with allometry)
```{r}
mean.PCA <- gm.prcomp(mean.shapes)

# Make axis labels
PCs.allo <- mean.PCA$pc.summary$importance
PCx.per.allo <- round(PCs.allo[2, 1] * 100, digits = 1)  # % with 1 decimal
PCx.lab.allo <- paste("PC1", " (", PCx.per.allo, "%)", sep = "")
PCy.per.allo <- round(PCs.allo[2, 2] * 100, digits = 1)
PCy.lab.allo <- paste("PC2", " (", PCy.per.allo, "%)", sep = "")
```

## B-D) PCAs of shape residual dataset (WITHOUT allometry)
All other PCA plots in figure use this dataset

### Extract shape residuals by removing allometric shape signal
Code thanks to Dr Vera Weisbecker
```{r}
# Use pgls of coords with size to get residuals
allom.GDF <- geomorph.data.frame(coords = mean.shapes, Csize = info.means$MeanCsize)
allometry.all <- procD.pgls(coords ~ Csize, aus.tree, data = allom.GDF)

# Turn the residuals into an appropriate array
dimnames(mean.shapes)[[3]] <- info.means$FullName
allom.resid <- arrayspecs(allometry.all$pgls.residuals, dim(mean.shapes)[[1]], 3)

# Add the consensus shape (from the GPA) to the residuals if you want to understand the landmark variation in the residuals - needed for script 06
gpa <- gpagen(mean.shapes)
residuals <- allom.resid + array(gpa$consensus, dim(allom.resid))
```

### Make PCA of shape residuals
Color and legend by species
```{r}
# PCA in part B and D
res.PCA <- gm.prcomp(residuals)

# Make axis labels
PCs <- res.PCA$pc.summary$importance
PCx.per <- round(PCs[2, 1] * 100, digits = 1)  # % with 1 decimal
PCx.lab <- paste("PC1", " (", PCx.per, "%)", sep = "")
PCy.per <- round(PCs[2, 2] * 100, digits = 1)
PCy.lab <- paste("PC2", " (", PCy.per, "%)", sep = "")
```

### C) Plot colors by diet and point shapes by locomotion
```{r}
# Set up palette for 5 diet types
warm <- sequential_hcl(5, "LaJolla")  
col.carn <- warm[4]  # meat colored red
col.frug <- warm[3]  # fruit colored orange
col.gran <- warm[2]  # grain colored gold
col.foli <- sequential_hcl(5, "Greens 3")[2]  # green grass color for folivores
col.omni <- sequential_hcl(5, "Grays")[3]  # gray for anything goes omnivores

# Define color palette for diets of mean specimens
col.diet.5 <- c(col.carn, col.frug, col.gran, col.foli, col.omni)
col.diet.means <- PlotByGroup(info.means, "Feeding", col.diet.5)

# Points for 4 locomotion types: Aquatic (circle), Arboreal (triangle), Terrestrial (square), hopping (diamond)
pch.gld.4 <- c(16, 24, 23, 22)
pch.gld.means <- PlotByGroup(info.means, "Guild", pch.gld.4)

# Genera legend 
genera.phylo.full <- c("Mus", "Rattus", "Pogonomys", "Xeromys", "Hydromys", "Notomys", "Pseudomys", "Zyzomys", "Leggadina", "Uromys", "Melomys", "Leporillus", "Mesembriomys", "Conilurus")
```

### D) Plot PCA of shape residuals without Notomys
The bipedal hopping locomotion of Notomys could be driving a signal not found in the remaining dataset. To test, we take these 4 species out. 
```{r}
# Remove Notomys species from dataset
index.not <- which(info.means$Genus == "Not")
residuals.no <- residuals[, , -index.not]
info.means.no <- info.means[-index.not, ]

# PCA
wrap.PCA.r.n <- gm.prcomp(residuals.no)
alpha.col.no <- alpha.col[-index.not]
alpha.pch.no <- alpha.pch[-index.not]

# Calulate % of PC
PCs.n <- wrap.PCA.r.n$pc.summary$importance
PCx.per.n <- round(PCs.n[2, 1] * 100, digits = 1)  # % with 1 decimal
PCx.lab.n <- paste("PC1", " (", PCx.per.n, "%)", sep = "")
PCy.per.n <- round(PCs.n[2, 2] * 100, digits = 1)
PCy.lab.n <- paste("PC2", " (", PCy.per.n, "%)", sep = "")
```

### Prepare evolutionary allometry information (from AmNat paper) for Fig 1
```{r}
# Run Procrustes ANOVA with interaction
pgls.gdf <- geomorph.data.frame(coords = mean.shapes, size = info.means$MeanCsize, clade = info.means$Clade)

fit.pgls <- procD.pgls(coords ~ log(size) * clade, aus.tree, iter = 500, data = pgls.gdf, print.progress = FALSE)

# Find Reg Score values for plotting purposes
evo.allo.rs <- plotAllometry(fit.pgls, size = info.means$MeanCsize, method = "RegScore")
```

### Export Figure 1 to pdf
Finally, we compare the full shape PCA to the shape residuals PCAs. All PCAs are relevant to Figure 2 heatmaps.
```{r}
# Set up Notomys abline
not.index <- which(info.means$Genus.y == "Notomys")
not.Reg <- evo.allo.rs$RegScore[not.index]
not.Cs <- info.means$MeanCsize[not.index]

# Set up for export and for multiple panels
part.coords <- c(-0.117, 0.095)  # plotting parameters for A, C, D, E placement

setEPS()
postscript("../Data/Results/Figure1_PCAs_6plot.eps")
mat <- matrix(c(1:6), 3, 2, byrow = TRUE)  # 6 plots, 3 rows, 2 columns
layout(mat, widths = rep(1, dim(mat)[2]), heights = rep(0.5, dim(mat)[1]))   

# A) FULL SHAPE PCA - diet & locomotion legend
par(mar = c(4.5, 4.5, 2, 1))
plot(x = -mean.PCA$x[, 1],  # Flip for consistency
                 y = mean.PCA$x[, 2],  
                 xlab = "PC1 (52.3%)", 
                 ylab = "PC2 (14.5%)",
                 xlim = c(-0.12, 0.12),
                 ylim = c(-0.1, 0.1),
                 col = col.diet.means, 
                 pch = pch.gld.means, 
                 bg = col.diet.means,
                 cex = 1.5,
                 cex.axis = 1.1, 
                 cex.lab = 1.3)
text(part.coords[1], part.coords[2], "a", cex = 1.5)
text(-0.045, 0.092, "full shape dataset (ecology)")
text(-0.078, 0.075, expression(paste("Marcy ", italic("et al. "), "2020")), cex = 0.8)

legend(-0.131, 0.067, legend = c("Arboreal", "Hopping", "Terrestrial", "Semiaquatic", "", "Carnivore", "Frugivore", "Granivore", "Folivore", "Omnivore"), col = c(rep("black", 4), "white", col.diet.5), border = NULL, pch = c(24:21, rep(16, 6)), cex = 0.8, bty = "n")

# B) Evolutionary Allometry plot - diet & locomotion legend
par(mar = c(4, 4, 1, 1))
plot(x = log(info.means$MeanCsize),
     y = -evo.allo.rs$RegScore,
     xlim = c(4.75, 6),
     col = col.diet.means, 
     pch = pch.gld.means, 
     bg = col.diet.means,
     main = " ",
     xlab = "Log centroid size", 
     ylab = "Shape  (Reg Score)",
     bty = "n")
text(4.8, 0.006, "b", cex = 1.5)
text(5.1, 0.0058, "evolutionary allometry")
abline(lm(-evo.allo.rs$RegScore ~ log(info.means$MeanCsize)), col = "dark grey", lwd = 1)  # common evolutionary allometry line
#text(x = log(info.means$MeanCsize), y = -evo.allo.rs$RegScore, labels = info.means$Genus)  # labels genera on plot points

# C) FULL SHAPE PCA - genus colors
par(mar = c(4.5, 4.5, 2, 1))
plot(x = -mean.PCA$x[, 1],  # flip for consistency
                 y = mean.PCA$x[, 2],  
                 xlab = "PC1 (52.3%)", 
                 ylab = "PC2 (14.5%)",
                 xlim = c(-0.12, 0.12),
                 ylim = c(-0.1, 0.1),
                 col = alpha.col,
                 pch = 16,
                 bg = alpha.col,
                 cex = 1.5,
                 cex.axis = 1.1, 
                 cex.lab = 1.3)
text(part.coords[1], part.coords[2], "c", cex = 1.5)
text(-0.046, 0.092, "full shape dataset (genera)")

legend("bottomleft", legend = genera.phylo.full, col = col.phylo, pch = 16, cex = 0.68, pt.cex = 0.85, bty = "n", text.font = 3)

# D) shape residuals PCA - genus colors
par(mar = c(4.5, 4.5, 2, 1))
plot(x = res.PCA$x[, 1],  # no flip for consistency
                 y = -res.PCA$x[, 2],  
                 xlab = "PC1 (26.8%)", 
                 ylab = "PC2 (18.6%)",
                 xlim = c(-0.12, 0.12),
                 ylim = c(-0.1, 0.1),
                 col = alpha.col,
                 pch = 16,
                 bg = alpha.col,
                 cex = 1.5,
                 cex.axis = 1.1, 
                 cex.lab = 1.3)
text(part.coords[1], part.coords[2], "d", cex = 1.5)
text(-0.035, 0.092, "shape residual dataset (genera)")

# E) No Notomys shape residuals PCA - genus colors
par(mar = c(4.5, 4.5, 2, 1))
plot(x = wrap.PCA.r.n$x[, 1],  
                 y = wrap.PCA.r.n$x[, 2],  # no flip for consistency
                 xlab = "PC1 (30.1%)", 
                 ylab = "PC1 (16.7%)",
                 xlim = c(-0.12, 0.12),
                 ylim = c(-0.1, 0.1),
                 col = alpha.col.no, 
                 pch = 16, 
                 bg = alpha.col.no,
                 cex = 1.5,
                 cex.axis = 1.1, 
                 cex.lab = 1.3)
text(part.coords[1], part.coords[2], "e", cex = 1.5)
text(-0.034, 0.092, expression(paste("shape residual without ", italic("Notomys"))))

# F) shape Residuals PCA - ecology
par(mar = c(4.5, 4.5, 2, 1))
plot(x = res.PCA$x[, 1],  # no flip for consistency
                 y = -res.PCA$x[, 2],  
                 xlab = "PC1 (26.8%)", 
                 ylab = "PC1 (18.6%)",
                 xlim = c(-0.12, 0.12),
                 ylim = c(-0.1, 0.1),
                 col = col.diet.means, 
                 pch = pch.gld.means, 
                 bg = col.diet.means,
                 cex = 1.5,
                 cex.axis = 1.1, 
                 cex.lab = 1.3)
text(part.coords[1], part.coords[2], "f", cex = 1.5)
text(-0.053, 0.092, "shape residual (ecology)")

dev.off()
```

## 2) Compare the full shape and shape residual datasets
Mantel tests to compare full shape morphospace to the shape residual morphospace
```{r}
#A quick check of the correlation betwen full shape PC2 and residual shape PC1:
cor(mean.PCA$x[,2], res.PCA$x[,1])


# Make distance matrices
d.full.pca <- dist(mean.PCA$x)
d.res.pca <- dist(res.PCA$x)

# Mantel test for all PC scores
man.all <- mantel(d.full.pca, d.res.pca, method = "pearson", permutations = 1000) 
man.all$statistic
man.all$signif
```

### Mantel tests to compare just the 3 most important PCs
```{r}
# Make distance matrices
d.full.pc <- dist(mean.PCA$x[, 1:3])
d.res.pc <- dist(res.PCA$x[, 1:3])

# Mantel test for all PC scores
man.pc13 <- mantel(d.full.pc, d.res.pc, method = "pearson", permutations = 1000) 
man.pc13$statistic
man.pc13$signif
```

### Comaparison of morphospaces without allometric PC1
```{r}
# Make distance matrices
d.full.wo1 <- dist(mean.PCA$x[, 2:dim(mean.PCA$x)[2]])
d.res.woL <- dist(res.PCA$x[, 1:dim(res.PCA$x)[2]-1])

# Mantel test for all PC scores
man.pc.wo1 <- mantel(d.full.wo1, d.res.woL, method = "pearson", permutations = 1000) 
man.pc.wo1$statistic
man.pc.wo1$signif
```


### Correlation of residual PC1 to full shape PC2 and vice versa
```{r}
cor.pc1 <- cor(mean.PCA$x[, 2], -res.PCA$x[, 1])
cor.pc1

cor.pc2 <- cor(mean.PCA$x[, 1], res.PCA$x[, 2])
cor.pc2

cor.pc2v3 <- cor(mean.PCA$x[, 3], res.PCA$x[, 2])
cor.pc2v3
```

### Correlation of full shape PC1 and residual PC2 to centroid size
```{r}
cor.size <- cor(-mean.PCA$x[, 1], info.means$MeanCsize)
cor.size

# Correlation of PC2 to centroid size
cor.pc2.size <- cor(res.PCA$x[, 2], info.means$MeanCsize)
cor.pc2.size
```


## 3) Export skreeplots for PC importance scores for 3 separate datasets
Set up 3 datasets for screeplots
```{r}
# Full shape
full.PC.impt <- data.frame(PC = 1:(length(summary(mean.PCA)$PC.summary)), impt = unlist(summary(mean.PCA)[[1]][2,]))

# Shape residual
res.PC.impt <- data.frame(PC = 1:(length(summary(res.PCA)$PC.summary)), impt = unlist(summary(res.PCA)[[1]][2,]))

# Shape residual without Notomys wrap.PCA.r.n
rnot.PC.impt <- data.frame(PC = 1:(length(summary(wrap.PCA.r.n)$PC.summary)), impt = unlist(summary(wrap.PCA.r.n)[[1]][2,]))

```

### Create supplementary figure 1
```{r}
# Make simple multipanel figure
pdf("../Data/Results/SFigure1_Screeplots.pdf")
mat <- matrix(c(1:3), 3, 1, byrow = TRUE)  # 3 plots, 3 rows, 1 column
layout(mat, widths = rep(0.5, dim(mat)[2]), heights = rep(0.25, dim(mat)[1])) 

# 1) 
barplot(full.PC.impt[, 2], ylim = c(0, 0.5))
abline(h = 0.1, col = 'deeppink4')
mtext("A) full shape dataset", 3, line = 1.5, cex = 0.9, adj = 0, col = "black")
mtext("% of variance", 2, line = 2.5, cex = 0.83, col = "black")

# 2)
barplot(res.PC.impt[, 2], ylim = c(0, 0.5))
abline(h = 0.1, col = 'deeppink4')
mtext("B) shape residual dataset", 3, line = 1, cex = 0.9, adj = 0, col = "black")
mtext("% of variance", 2, line = 2.5, cex = 0.83, col = "black")

# 3)
barplot(rnot.PC.impt[, 2], ylim = c(0, 0.5))
abline(h = 0.1, col = 'deeppink4')
mtext(expression(paste("C) shape residual dataset without ", italic(Notomys))), 3, line = 1, cex = 0.9, adj = 0, col = "black")
mtext("% of variance", 2, line = 2.5, cex = 0.83, col = "black")
mtext("Principal component", 1, line = 1, cex = 0.83, col = "black")

dev.off()
```

### Save intermediate data
```{r}
save(mean.PCA, res.PCA, allom.resid, residuals, file = "../Data/Processed/04-residual-data.rda")
```
