==> Analysis/04-plot-exploratory-PCAs.Rmd <==
---
title: "04-plot-exploratory-PCAs"
author: "Ariel Marcy"
date: "2018-10-17"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = '../eco-rodents')
```

# Plotting exploratory PCAs
PCAs provide a human-friendly way to visualize shape data. They are a good way to understand what's going on in the data quickly. However, because they artificially collapse the shape variation into orthogonal PC axes, they must be considered exploratory analyses only.

Here, we explore the variation in rodent skulls by coloring the PCA plots by genus, by traits, and by biome region.

### Load packages, functions, and data
```{r message = FALSE}
library(stringr)
library(data.table)
library(geomorph)
library(colorspace)  # colors
source("../Functions/utilities.R")  # custom functions
load(file = "../Data/Processed/03-main-data.rda")
load(file = "../Data/Processed/03-graphing-vectors.rda")
```

## Plot PCA by genus for small patch dataset
The function `PlotPCA()` quickly plots a PCA with specimens colored by groups and optionally given different point shapes for other attributes. The function also reports PC axis variation in %. 
```{r}
PCA <- plotTangentSpace(shape)

# Plot PCA by genus and export as .eps into ../Data/Results folder
setEPS()
postscript("../Data/Results/PCA_genus.eps")

PlotPCA(PCA, 1, 2, sp.col.gen, pch.spp)
legend("bottomright", legend = phylo.names, col = col.unique.spp, pch = pch.unique.spp, pt.bg = col.unique.spp, cex = .35, pt.cex = .7, ncol = 2)

dev.off()
```

## PCA plot by ancestor group
There are two main tribes on Australia: Hydromini and Rattini, which represent two different lineages within Murinae to make it from Asia to Sahul. 

Within Hydromini, there are 5 groups with unique ancestors and arrivals from New Guinea to Australia. Most had unique radiations that centered on different biomes. 
```{r}
# Set up colors, using distinct colors from above. 
col.Hyd <- col.unique.spp[which(phylo.names == "H chr")]  # Hydromys 
col.invR <- "black"  # to distinguish between native and invasive Rattus rats 
col.Mes <- col.unique.spp[which(phylo.names == "M gou")]  # Mesembriomys
col.mus <- col.unique.spp[which(phylo.names == "M mus")]  # Mus musculus
col.Pog <- col.unique.spp[which(phylo.names == "P mol")]  # Pogonomys
col.Pse <- col.unique.spp[which(phylo.names == "L for")]  # Pseudomys
col.Rat <- col.unique.spp[which(phylo.names == "R vil")]  # native Rattus
col.Uro <- col.unique.spp[which(phylo.names == "M bur")]  # Uromys

col.clade <- c(col.Hyd, col.invR, col.Mes, col.mus, col.Pog, col.Pse, col.Rat, col.Uro)

# Colors matching each specimen in order of shape dataset
col.by.clade <- PlotByGroup(info, "Clade", col.clade)

# PCA 
PlotPCA(PCA, 1, 2, col.by.clade)
```

## PCA plot by diet and guild
This PCA colors specimens by diet and assigns point shape based on guild, i.e. where they spend most of their time.
```{r}
# Set up pallete for diet
warm <- sequential_hcl(5, "LaJolla")  
col.carn <- warm[4]  # meat colored red
col.frug <- warm[3]  # fruit colored orange
col.gran <- warm[2]  # grain colored gold
col.herb <- sequential_hcl(5, "Greens 3")[2]  # green colored grass
col.omni <- sequential_hcl(5, "Grays")[3]  # gray for anything goes

# Define color pallete for diets of all specimens
col.diet <- c(col.carn, col.frug, col.gran, col.herb, col.omni)
col.diet <- PlotByGroup(info, "Feeding", col.diet)

# Points: Aquatic (circle), Arboreal (triangle), Terrestrial (square)
pch.gld <- c(19, 24, 22)
pch.gld <- PlotByGroup(info, "Guild", pch.gld)
```

Plot PC1 vs PC2 and add custom legends.
```{r}
PlotPCA(PCA, 1, 2, col.diet, pch.gld)
legend("bottomright", legend = c("Aquatic", "Arboreal", "Terrest."), col = "black", border = NULL, pch = c(21, 24, 22))
text(0.02, -0.09, "Carnivore", col = col.carn)
text(0.11, -0.02, "Frugivore", col = col.frug)
text(0.08, 0.045, "Herbivore", col = col.herb)
text(-0.06, -0.05, "Omnivore", col = col.omni)
text(-0.11, 0.01, "Granivore", col = col.gran)
```

## Morphospace partitioning by habitat
So far, most of the variation in the morphospace could be explained by size and genus affiliation, with different genera of similar sizes overlapping quite a bit. There doesn't seem to be much differentiation among Australian rodents, but perhaps the differentiation is hidden by geographical region or environmental partitioning. 

### Write functions to visualize rodent morphospace by region
Here, we subset the PCA scores by species which inhabit the 7 different ecological regions of Australia. Regions and species distributions therein taken from Breed & Ford 2007. `FoundInRegion()` subsets existing PCA data by species present in given region(s).

We'll use a loop to plot PCAs for each region, with unique species within each genera having a unique point shape. The legends only name the species present for a given region. 
```{r}
# Set up loop for particular regions in Australia
region.names <- c("Savannah", "Arid Zone", "NE Wet Forest", "NE Dry Forest", "Pilbara", "South West", "South East")
pc.scores <- subset(PCA$pc.scores, select = 1:3)  # only 1st 3 PCs

# Write x and y labels with proportion of variance for PCx and PCy
PCs <- PCA$pc.summary$importance
PCx.per <- round(PCs[2, 1] * 100, digits = 1)  # % with 1 decimal
PCx.lab <- paste("PC1 (", PCx.per, "%)", sep = "")
PCy.per <- round(PCs[2, 2] * 100, digits = 1)
PCy.lab <- paste("PC2 (", PCy.per, "%)", sep = "")


par(mfrow=c(4,2), mar = c(2,2,2,2))
# Loop through all 7 regions to plot morphospace of species found in each
for (i in 1:7) {
        # Subset info.traits and PCA by region
        in.region <- FoundInRegion(info, regions = i)
        info.region <- info[in.region, ]
        pc.region <- as.data.frame(pc.scores[in.region, ])
        
        # Subset plotting colors and pch symbols for spp in region
        col.taxa.region <- sp.col.gen[in.region]  # colors for all spp
        pch.taxa.region <- pch.spp[in.region]  # pch for all spp
        
        # Plot PCA with legend and labels
        PCA.plot <- plot(x = pc.region$PC1,
                         y = pc.region$PC2,
                         main = paste(region.names[i], "Region Morphospace" ),
                         xlab = PCx.lab,
                         ylab = PCy.lab,
                         xlim = c(-0.11, 0.18),
                         ylim = c(-0.095, 0.05),
                         col = col.taxa.region, 
                         pch = pch.taxa.region, 
                         bg = col.taxa.region,
                         cex = 1.5,
                         cex.axis = 1.3, 
                         cex.lab = 1.3)
}
```

## Remove specialists from PCAs
This step is to see whether the PCA changes to highlight morphological differences among the less specialized species.

### No aquatic carnivores
These are the most specialized so most likely to skew the visualization / hide variation in other areas of the PCA
```{r}
test.data <- info[, c("Taxa", "Genus", "Species", "CatNum", "Order", "Csize", "Clade", "Feeding", "Guild")]  # relevant columns

# Subset out Hydromys and Xeromys
aquatic <- which(info$Guild == "Aquatic")  # specimens to remove
test.data <- test.data[-aquatic, ]  # metadata
test.shape <- shape[, , -aquatic]  # shape data
test.col.gen <- sp.col.gen[-aquatic]  # color information
test.pch.gen <- pch.spp[-aquatic]  # point shape information

# Run PCA
no.aq.pca <- plotTangentSpace(test.shape, groups = test.col.gen, axis1 = 1, axis2 = 2, verbose = T)

# Plot PCA
PlotPCA(no.aq.pca, 1, 2, test.col.gen, test.pch.gen)
```

### Aquatic carnivores only
These are the most specialized so most likely to skew the visualization / hide variation in other areas of the PCA
```{r}
test.data <- info.traits[, c("Taxa", "Genus", "Species", "CatNum", "Order", "Csize", "Clade", "Feeding", "Guild")]  # relevant columns

# Subset out Hydromys and Xeromys
aquatic <- which(info.traits$Guild == "Aquatic" | info.traits$Feeding == "Frugivorous")  # specimens to remove
test.data <- test.data[-aquatic, ]  # metadata
test.shape <- shape[, , -aquatic]  # shape data
test.col.gen <- sp.col.gen[-aquatic]  # color information
test.pch.gen <- sp.pch.taxa[-aquatic]  # point shape information

# Run PCA
no.aq.pca <- plotTangentSpace(test.shape, groups = test.col.gen, axis1 = 1, axis2 = 2, verbose = T)

# Plot PCA
PlotPCA(no.aq.pca, 1, 2, test.col.gen, test.pch.gen)
```

### Save intermediate data
```{r}
save(PCA, file = "../Data/Processed/04-PCA.rda")
```
