==> Analysis/04_02_Allometry_and_residuals.Rmd <==
---
title: "04-compare_PCAs"
author: "Ariel Marcy & Vera Weisbecker"
date: "8/15/2019"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE)
knitr::opts_knit$set(root.dir = '../eco-rodents')
```

# Plot PCAs for allometric and shape residual datasets
Plot PCAs for allometric shape and shape residual, provide coloration and pch symbols by locomotion and diet, and visualize changes in shape residual PCA when *Notomys* species are removed. These plots are relevant to the heatplots created in script 05 / shown in Figure 2. 

Creates multipanel Figure 1.

Compares morphospace similarity of the two datasets with Mantel tests and a correlation test.

Creates supplementary Figure 1, a screeplot of the the PCs for each dataset.

Produces shape residual data used in later analyses, stored in rda file 04.

### Load packages, functions, and data
Tree, mean shape, and metadata for phylogenetic analyses come from the allometry rodents script 04. 
```{r message = FALSE}
library(geomorph)   
library(vegan)      
library(dplyr)      
library(colorspace) 
library(phytools) 
library(geiger)      
library(mvMORPH)  
library(qpcR)


# This sets the wd to local
library(rstudioapi) # Safely Access the RStudio API
setwd(dirname(getActiveDocumentContext()$path))

source("../Functions/utilities.R")  # custom functions
load(file = "../Data/Processed/03-main-data.rda")
load(file = "../Data/Processed/03-tree-data.rda")
load(file = "../Data/Processed/03-graphing-vectors.rda")
load(file="../Data/Processed/allometry_residuals.rda")


#This is repeated in the modularity code but it's easier to just add it again here - less complicated.


lm.modules <- read.csv("../Data/Processed/Landmark_modules.csv", sep = ",", header = TRUE)

basicran <- mean.shapes[which(lm.modules[, 4] == "basicran"), ,]
molar <- mean.shapes[which(lm.modules[, 4] == "molar"), ,]
orbital <- mean.shapes[which(lm.modules[, 4] == "orbital"), ,]
rostrum <- mean.shapes[which(lm.modules[, 4] == "rostrum"), ,]
vault <- mean.shapes[which(lm.modules[, 4] == "vault"), ,]

#"Module_list" is modules from the same gpa; mean_modules is separate GPAs

module_list <- list( basicran=basicran, molar=molar, orbital=orbital, rostrum=rostrum, vault=vault)


basicran_res <- residuals[which(lm.modules[, 4] == "basicran"), ,]
molar_res <- residuals[which(lm.modules[, 4] == "molar"), ,]
orbital_res <- residuals[which(lm.modules[, 4] == "orbital"), ,]
rostrum_res <- residuals[which(lm.modules[, 4] == "rostrum"), ,]
vault_res <- residuals[which(lm.modules[, 4] == "vault"), ,]

module_list_res <- list( basicran_res=basicran_res, molar_res=molar_res, orbital_res=orbital_res, rostrum_res=rostrum_res, vault_res=vault_res)




```

#Phylogenetic signal of shape and size
```{r}


Physig_allshape <- physignal(mean.shapes, aus.tree )


#pull out csize so it also has row names and check

log_mean_csize <- log(info.means$MeanCsize)
names(log_mean_csize) <- info.means$FullName
names(log_mean_csize) == aus.tree$tip.label

physignal(log_mean_csize, aus.tree )


#now csizes of separate modules

physig_table_modules_csize <- matrix(ncol = 2, nrow= length(module_list))

rownames(physig_table_modules_csize)<- names(mean_modules)
colnames(physig_table_modules_csize) <- c("K", "p")

#as it also says in the help file, phylosig from phytools and physignal from geomorph give the same values in univariate data.

for (i in (1: length(mean_modules))){
  
  temp <- physignal(mean_modules[[i]]$csize, aus.tree)
  
  physig_table_modules_csize[i,1] <- temp$phy.signal
   physig_table_modules_csize[i,2] <- temp$pvalue

  
}

write.csv(physig_table_modules_csize, file="../Data/Results/physig_table_modules_csize.csv")

```

#Now we want to know allometry in separately GPA-ed modules

```{r}

#add empty colums for values
Allometry <- matrix(ncol=7, nrow=length(names(mean_modules)), dimnames=list(c(names(mean_modules)), c("Df", "SS","MS" ,"Rsq", "F", "Z","P")) )


for (i in 1:length(mean_modules)){
  
temp <- procD.pgls( mean_modules[[i]]$shape ~ log(mean_modules[[i]]$csize), phy=aus.tree,lambda = 1, iter = 10000)

Allometry[i, 1:7] <- unlist( temp$aov.table[1,])

}




Allometry_non_phylo <- matrix(ncol=7, nrow=length(names(mean_modules)), dimnames=list(c(names(mean_modules)), c("Df", "SS","MS" ,"Rsq", "F", "Z","P")) )


for (i in 1:length(mean_modules)){
  
temp <- procD.lm( mean_modules[[i]]$shape ~ log(mean_modules[[i]]$csize), iter = 10000)

Allometry_non_phylo [i, 1:7] <- unlist( temp$aov.table[1,])

}


Allometry_gls_lm <- rbind(Allometry, c(rep(NA, 7)), Allometry_non_phylo)


#Take out df column

Allometry_gls_lm <- Allometry_gls_lm[,-1]
#round SS/MS
Allometry_gls_lm[,c(1,2,6)] <- round (Allometry_gls_lm[,c(2,3,6)], 3)
Allometry_gls_lm[,c(3:5)] <- round (Allometry_gls_lm[,c(3:5)], 2)

write.csv(Allometry_gls_lm, file="../Data/Results/Module_allometry.csv")

#it is also worth checking whether this pattern holds up if we use the overall csize of the skull, since the datasets are all aligned.

#add empty colums for values
Allometry_fullcsize <- matrix(ncol=7, nrow=length(names(mean_modules)), dimnames=list(c(names(mean_modules)), c("Df", "SS","MS" ,"Rsq", "F", "Z","P")) )


for (i in 1:length(mean_modules)){

temp <- procD.pgls( mean_modules[[i]]$shape ~ log(info.means$MeanCsize), phy=aus.tree)

Allometry_fullcsize[i, 1:7] <- unlist( temp$aov.table[1,])

}


#The result is not as expected. so just plotting a check here

mfrow3d(2,3)

for (i in (1:length(mean_modules))){


  spheres3d(mean_modules[[i]]$shape[,,1], radius=0.01, col="hotpink", )
 title3d(main=names(mean_modules)[i])
 next3d()

}



```

#Does this pan out differently if there isn't a joint superimposition?

```{r}

#add empty colums for values
Allometry_joint_gpa <- matrix(ncol=7, nrow=length(names(mean_modules)), dimnames=list(c(names(mean_modules)), c("Df", "SS","MS" ,"Rsq", "F", "Z","P")) )


for (i in 1:length(module_list)){
  
temp <- procD.pgls( module_list[[i]] ~ log(info.means$MeanCsize), phy=aus.tree, iter = 10000)

Allometry_joint_gpa[i, 1:7] <- unlist( temp$aov.table[1,])

}

#



Allometry_joint_gpa_non_phylo <- matrix(ncol=7, nrow=length(names(mean_modules)), dimnames=list(c(names(mean_modules)), c("Df", "SS","MS" ,"Rsq", "F", "Z","P")) )


for (i in 1:length(mean_modules)){
  
temp <- procD.lm( module_list[[i]] ~ log(info.means$MeanCsize), iter = 10000)

Allometry_joint_gpa_non_phylo [i, 1:7] <- unlist( temp$aov.table[1,])

}


Allometry_gls_lm_joint_gpa <- rbind(Allometry_joint_gpa, c(rep(NA, 7)), Allometry_joint_gpa_non_phylo)


#Take out df column

Allometry_gls_lm_joint_gpa <- Allometry_gls_lm_joint_gpa[,-1]
#round SS/MS
Allometry_gls_lm_joint_gpa[,c(1,2,6)] <- round (Allometry_gls_lm_joint_gpa[,c(2,3,6)], 3)
Allometry_gls_lm_joint_gpa[,c(3:5)] <- round (Allometry_gls_lm_joint_gpa[,c(3:5)], 2)



write.csv(Allometry_gls_lm_joint_gpa, file="../Data/Results/Module_allometry_joint_gpa.csv")

#Here making a table of joint vs. separate gpa instead

Allometry_separate_vs_joint <- rbind(Allometry_joint_gpa, c(rep(NA, 7)), Allometry)

#Take out df column

Allometry_separate_vs_joint <- Allometry_separate_vs_joint[,-1]
#round SS/MS
Allometry_separate_vs_joint[,c(1,2,6)] <- round (Allometry_separate_vs_joint[,c(2,3,6)], 3)
Allometry_separate_vs_joint[,c(3:5)] <- round (Allometry_separate_vs_joint[,c(3:5)], 2)

write.csv(Allometry_separate_vs_joint, file="../Data/Results/Allometry_separate_vs_joint.csv")

```

#Strange that the rsquareds for the rostrum are so low. How about disparity?

```{r}

disparity_table <- matrix (ncol=3, nrow=length(module_list))
colnames(disparity_table) <- c("Full dataset", "Residuals", "% reduction")
rownames(disparity_table)<- names(module_list)

for (i in (1: length(module_list))){
  
  temp <- round ((morphol.disparity(module_list[[i]]~1)/length(module_list_res[[i]][,,1][,1]))*10000, 3)
  
  disparity_table[i,1] <- temp
   

  
}

for (i in (1: length(module_list_res))){
  
  temp <- round((morphol.disparity(module_list_res[[i]]~1)/length(module_list_res[[i]][,,1][,1]))*10000, 3)
  
  disparity_table[i,2] <- temp
   

  
}

#By how much does disparity drop when you take allometry out?

disparity_table[,3]<- round (((disparity_table[,1]-disparity_table[,2])*100)/(disparity_table[,1]),1)

disparity_table


write.csv(disparity_table, file="../Data/Results/disparity_tablex10000.csv")

```
