==> Analysis/05-test-partitioning.Rmd <==
---
title: "05-test-partitioning"
author: "Ariel Marcy & Vera Weisbecker"
date: "8/15/2019"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = '../eco-rodents')
```

# Test biome partitioning 
First, testing for a phylogenetic signal difference in each of the 7 regional biomes. 

Second, testing for partitioning by region in the residual shapes (size-less shape)

### Load packages, functions, and data
Tree, mean shape, and metadata for phylogenetic analyses come from the allometry rodents script 04. 
```{r message = FALSE}
library(geomorph)
library(dplyr)
library(geiger)
source("../Functions/utilities.R")  # custom functions
load(file = "../Data/Processed/03-main-data.rda")
load(file = "../Data/Processed/03-tree-data.rda")
load(file = "../Data/Processed/03-graphing-vectors.rda")
```

## Assess phylogenetic signal overall
Baseline comparison.
```{r}
# Test phylogenetic signal
physig.all <- physignal(mean.shapes, aus.tree)
summary(physig.all)
```

## 1) Assess phylogenetic signal in each region partition
Here, again, we subset the PCA scores by species which inhabit the 7 different ecological regions of Australia. We'll use a loop to test the phylogenetic signal in each region. 

Evolutionary tree comes from the Allometry Paper repo. 
```{r}
traits <- info[, -c(4, 5, 6)]  # remove columns that ID individuals of species

info.means$Order <- c(1:nrow(info.means))  # preserves order with coordinate shape data

# Set up loop for particular regions in Australia
region.names <- c("Savannah", "Arid Zone", "NE Wet Forest", "NE Dry Forest", "Pilbara", "South West", "South East")
n.taxa <- c(rep("NA", length(region.names)))
K.region <- c(rep("NA", length(region.names)))
p.region <- c(rep("NA", length(region.names)))

# Loop through all 7 regions to plot morphospace of species found in each
for (i in 1:7) {
        # Identify species found in the region
        spp.in.region <- FoundInRegion(traits, regions = i)
        spp.region <- traits[spp.in.region, ]
        
        # Subset metadata by region
        temp <- info.means[, c("Taxa", "MeanCsize", "FullName", "Order")]
        meta.region <- unique(merge(spp.region, temp, by = "FullName"))
        
        # Subset shape coordinates by region
        shape.region <- mean.shapes[, , meta.region$Order]
        
        # Subset tree by species in the region
        concord <- name.check(aus.tree, data.names = meta.region$FullName)  # check concordance
        region.tree <- drop.tip(aus.tree, concord$tree_not_data)  # prune tree to regional sample
        n.taxa[i] <- length(region.tree$tip.label)
        
        # Re-order to match tree tip order
        tree.order <- match(region.tree$tip.label, meta.region$FullName)
        meta.region <- meta.region[tree.order, ]  # metadata
        shape.region <- shape.region[, , tree.order]  # shape data
        
        # Test phylogenetic signal
        physig.region <- physignal(shape.region, region.tree)
        K.region[i] <- round(physig.region$phy.signal, 3)
        p.region[i] <- as.numeric(physig.region$pvalue)  # round after Bonferroni correction in next code chuck
}
```

### Make and export Table TK
Create and export Table TK. Includes Bonferroni p-value adjustment for running 7 biomes. 
```{r}
p.region <- round(p.adjust(p.region, method = "bonferroni"), 3)

table <- cbind(n.taxa, K.region, p.region)
colnames(table) <- c("N Taxa", "Phy. Signal", "p Value (cor.)")
row.names(table) <- region.names

write.table(table, "../Data/Results/Table_RegionPhySig.csv", sep = ",", col.names = NA)
```

## 2) Do same analysis but for centroid size only
```{r}
# Set up blank variables again for results table
n.taxa <- c(rep("NA", length(region.names)))
K.region <- c(rep("NA", length(region.names)))
p.region <- c(rep("NA", length(region.names)))

# Loop through all 7 regions to plot morphospace of species found in each
for (i in 1:7) {
        # Identify species found in the region
        spp.in.region <- FoundInRegion(traits, regions = i)
        spp.region <- traits[spp.in.region, ]
        
        # Subset metadata by region
        temp <- info.means[, c("Taxa", "MeanCsize", "FullName", "Order")]
        meta.region <- unique(merge(spp.region, temp, by = "FullName"))
        
        # Subset tree by species in the region
        concord <- name.check(aus.tree, data.names = meta.region$FullName)  # check concordance
        region.tree <- drop.tip(aus.tree, concord$tree_not_data)  # prune tree to regional sample
        n.taxa[i] <- length(region.tree$tip.label)
        
        # Re-order to match tree tip order
        tree.order <- match(region.tree$tip.label, meta.region$FullName)
        meta.region <- meta.region[tree.order, ]  # metadata
        
        # Test phylogenetic signal with centroid size
        Csize <- as.numeric(meta.region$MeanCsize)
        names(Csize) <- meta.region$FullName
        physig.region <- physignal(Csize, region.tree) 
        K.region[i] <- round(physig.region$phy.signal, 3)
        p.region[i] <- as.numeric(physig.region$pvalue)  # round after Bonferroni correction in next code chuck
}
```

### Make and export Table TK
Create and export Table TK. Includes Bonferroni p-value adjustment for running 7 biomes and filename indicating these results are for centroid size. 
```{r}
p.region <- round(p.adjust(p.region, method = "bonferroni"), 3)

table <- cbind(n.taxa, K.region, p.region)
colnames(table) <- c("N Taxa", "Phy. Signal", "p Value (cor.)")
row.names(table) <- region.names

write.table(table, "../Data/Results/Table_RegionPhySig_Csize.csv", sep = ",", col.names = NA)
```

## PCA with mean shape WITH ALLOMETRY for comparison
```{r}
alpha.col <- col.unique.spp[order(phylo.names)]
alpha.pch <- pch.unique.spp[order(phylo.names)]

mean.pca <- plotTangentSpace(mean.shapes, axis1 = 1, axis2 = 2, verbose = T)
PlotPCA(mean.pca, 1, 2, alpha.col, alpha.pch, flip.axis1 = T)
```

## 3) Check biome partitioning WITHOUT allometry and WITH phylogenetic correction
First, extract residuals - code thanks to Dr Vera Weisbecker
```{r}
# Use pgls of coords with size to get residuals
allom.GDF <- geomorph.data.frame(coords = mean.shapes, Csize = info.means$MeanCsize) 

dimnames(mean.shapes)[[3]] <- info.means$FullName

allometry.all <- procD.pgls(coords ~ Csize, aus.tree, data = allom.GDF)
summary(allometry.all)

# Turn the residuals into an appropriate array
allom.resid <- arrayspecs(allometry.all$pgls.residuals, dim(mean.shapes)[[1]], 3)

# Add the consensus shape (from the GPA) to the residuals if you want to understand the landmark variation in the residuals - needed for script 06
Allom.all.plus.consensus <- allom.resid + array(mean.shapes, dim(allom.resid))
```

### Then, plot residuals with consensus (WITHOUT allometry)
```{r}
res.PCA <- plotTangentSpace(Allom.all.plus.consensus, groups = alpha.col)

PlotPCA(res.PCA, 1, 2, alpha.col, alpha.pch, flip.axis1 = TRUE, flip.axis2 = TRUE)
legend("bottomright", phylo.names, col = col.unique.spp, pch = pch.unique.spp, pt.bg = col.unique.spp, cex = .53, pt.cex = .7, ncol = 2)
```

### Alternately, as above but with only residuals no consensus added
```{r}
res.PCA <- plotTangentSpace(allom.resid, groups = alpha.col)
PlotPCA(res.PCA, 1, 2, alpha.col, alpha.pch, flip.axis2 = TRUE)
legend("bottomright", phylo.names, col = col.unique.spp, pch = pch.unique.spp, pt.bg = col.unique.spp, cex = .53, pt.cex = .7, ncol = 2)
```

### Finally, plot residuals by region
```{r}
# Set up loop for particular regions in Australia
region.names <- c("Savannah", "Arid Zone", "NE Wet Forest", "NE Dry Forest", "Pilbara", "South West", "South East")
res.pc.scores <- subset(res.PCA$pc.scores, select = 1:3)  # only 1st 3 PCs

# Write x and y labels with proportion of variance for PCx and PCy
PCs <- res.PCA$pc.summary$importance
PCx.per <- round(PCs[2, 1] * 100, digits = 1)  # % with 1 decimal
PCx.lab <- paste("PC1 (", PCx.per, "%)", sep = "")
PCy.per <- round(PCs[2, 2] * 100, digits = 1)
PCy.lab <- paste("PC2 (", PCy.per, "%)", sep = "")

par(mfrow = c(4, 2), mar = c(2, 2, 2, 2))
# Loop through all 7 regions to plot morphospace of species found in each
for (i in 1:7) {
        # Subset info.traits and PCA by region
        spp.in.region <- FoundInRegion(traits, regions = i)
        spp.region <- traits[spp.in.region, ]
        
        # Subset metadata by region
        temp <- info.means[, c("Taxa", "MeanCsize", "FullName", "Order")]
        meta.region <- unique(merge(spp.region, temp, by = "FullName"))
        
        # Subset PCA coordinates by region
        res.pc.region <- as.data.frame(res.pc.scores[meta.region$Order, ])
        
        # Subset plotting colors and pch symbols for spp in region
        col.taxa.region <- alpha.col[meta.region$Order]  # colors for all spp
        pch.taxa.region <- alpha.pch[meta.region$Order]  # pch for all spp
        
        # Plot PCA with legend and labels
        PCA.plot <- plot(x = res.pc.region$PC1,
                         y = res.pc.region$PC2,
                         main = paste(region.names[i], "Residual Morphospace" ),
                         xlab = PCx.lab,
                         ylab = PCy.lab,
                         xlim = c(-0.12, 0.15),
                         ylim = c(-0.1, 0.2),
                         col = col.taxa.region, 
                         pch = pch.taxa.region, 
                         bg = col.taxa.region,
                         cex = 1.5,
                         cex.axis = 1.3, 
                         cex.lab = 1.3)
}
```

### Save intermediate data
```{r}
residuals <- Allom.all.plus.consensus  # shape data
save(alpha.col, alpha.pch, res.PCA, allom.resid, residuals, file = "../Data/Processed/05-residual-data.rda")
```
