==> Analysis/06-visualize-residual-morphospace.Rmd <==
---
title: "06-visualize-residual-morphospace"
author: "Ariel Marcy"
date: "9/18/2019"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = '../eco-rodents')
```

# Test biome partitioning 
1) Explore morphospace differences in PCA of size-less shape (residuals)

2) Assess patterns in skull shape variation along PC1 & 2 axes that could be related to adaptation

### Load packages, functions, and data
Tree, mean shape, and metadata for phylogenetic analyses come from the allometry rodents script 04. 
```{r message = FALSE}
library(geomorph)
library(data.table)
library(vegan)
library(geiger)  # for phylogenies
library(png)  # needed to work with screenshots
library(colorspace)  # plot colors

# Load libraries (and install if necessary) for landvR package
if(!require(devtools)) install.packages("devtools")
if(!require(dispRity)) install.packages("dispRity")
if(!require(landvR)) install_github("TGuillerme/landvR")
source("../Functions/utilities.R")  # custom functions
source("../Functions/plotGMPhyloMorphoSpace_plotmod.R")  # modified geomorph function
load(file = "../Data/Processed/03-main-data.rda")
load(file = "../Data/Processed/03-tree-data.rda")
load(file = "../Data/Processed/03-graphing-vectors.rda")
load(file = "../Data/Processed/05-residual-data.rda")
```

### Plot residuals - IMPORTANT FIGURE
Color and legend by species
```{r}
alpha.names <- dimnames(mean.shapes)[[3]]

res.PCA <- plotTangentSpace(residuals, groups = alpha.col, label = alpha.names)

# Make axis labels
PCs <- res.PCA$pc.summary$importance
PCx.per <- round(PCs[2, 1] * 100, digits = 1)  # % with 1 decimal
PCx.lab <- paste("PC1", " (", PCx.per, "%)", sep = "")
PCy.per <- round(PCs[2, 2] * 100, digits = 1)
PCy.lab <- paste("PC2", " (", PCy.per, "%)", sep = "")
        
# Pretty plot
PCA.plot <- plot(x = -res.PCA$pc.scores[, 1],  # Flip for consistency
                 y = -res.PCA$pc.scores[, 2],  
                 xlab = PCx.lab, 
                 ylab = PCy.lab,
                 asp = TRUE,
                 xlim = c(-0.12, 0.18),
                 ylim = c(-0.18, 0.12),
                 col = alpha.col, 
                 pch = alpha.pch, 
                 bg = alpha.col,
                 cex = 1.5,
                 cex.axis = 1.3, 
                 cex.lab = 1.3)
legend("bottomright", phylo.names, col = col.unique.spp, pch = pch.unique.spp, pt.bg = col.unique.spp, cex = .57, pt.cex = .7, ncol = 2)
```

## Plot warp grids of small versus large specimens
Uses Dr Thomas Guillerme's new package, `landvR` to see how landmarks vary within and between the two protocols.
```{r}
# Find mean configuration - i.e. mean shape in dataset
all.consensus <- (select.procrustes(residuals, selector = mean))[[1]]

# Measure spherical coordinates differences from the mean for each specimen for each landmark
diff.from.mean <- coordinates.difference(coordinates = residuals, reference = all.consensus, type = "spherical")

# Find largest and smallest native species from mean centroid size
big.index <- which(info.means$MeanCsize == max(info.means$MeanCsize))
biggest.shape <- mean.shapes[, , big.index]
sm.index <- which(info.means$Species == "del")
smallest.shape <- mean.shapes[, , sm.index]
```

# Plot the variation in residual shape change
Export images for plots
```{r}
# Dorsal view of median to biggest
open3d(windowRect = c(0, 0, 600, 600))
procrustes.var.plot(all.consensus, biggest.shape, col = heat.colors, col.val = diff.from.mean$Uromys_caudimaculatus[, "radius"])
view3d(phi = 0)  # flip cranium
rgl.snapshot("../Data/Processed/heatmap_mean-big_dor.png")

# Lateral view of median to biggest
open3d(windowRect = c(0, 0, 600, 600))
procrustes.var.plot(all.consensus, biggest.shape, col = heat.colors, col.val = diff.from.mean$Uromys_caudimaculatus[, "radius"])
view3d(phi = 90)
rgl.snapshot("../Data/Processed/heatmap_mean-big_lat.png")


# Dorsal view of median to smallest
open3d(windowRect = c(0, 0, 600, 600))
procrustes.var.plot(smallest.shape, all.consensus, col = heat.colors, col.val = diff.from.mean$Pseudomys_delicatulus[, "radius"]) 
view3d(phi = 0)  # flip cranium
rgl.snapshot("../Data/Processed/heatmap_mean-sm_dor.png")

# Lateral view of median to smallest
open3d(windowRect = c(0, 0, 600, 600))
procrustes.var.plot(smallest.shape, all.consensus, col = heat.colors, col.val = diff.from.mean$Pseudomys_delicatulus[, "radius"])
view3d(phi = 90)
rgl.snapshot("../Data/Processed/heatmap_mean-sm_lat.png")

# Remove whitespace (requires ImageMagik)
system("mogrify ../Data/Processed/heatmap_*.png -trim ../Data/Processed/*.png")
```

### Make quick figure comparing PC1 of with allometry vs without (residuals)
NOTE: this needs to be reproducible. Currently uses PNGs from allometry rodent side
```{r}
# Load in images
morph.dor.1 <- readPNG("../Data/Processed/PC1heatmap_dor.png")
morph.lat.1 <- readPNG("../Data/Processed/PC1heatmap_lat.png")

system("mogrify ../Data/Processed/res_heatmap_PC1_*.png -trim ../Data/Processed/*.png")  # removes whitespace (requires ImageMagik)
res.dor.1 <- readPNG("../Data/Processed/res_heatmap_PC1_dor.png")
res.lat.1 <- readPNG("../Data/Processed/res_heatmap_PC1_lat.png")

# Set up muli-panel figure
mat <- matrix(c(1, 2, 3, 4), 2, 2, byrow = TRUE)  # 2 rows, 2 columns
layout(mat, widths = c(1, 1, 1), heights = c(0.65, 0.75, 0.5))  

# 1) With Allometry: PC1 lateral view
par(mar = c(0, 0, 1, 0))  # sets the margins
plot(c(0, dim(morph.dor.1)[2]), c(0, dim(morph.dor.1)[1]), type = "n", axes = F, xlab = "", ylab = "", asp = T) 
rasterImage(morph.dor.1, 0, 0, dim(morph.dor.1)[2], dim(morph.dor.1)[1])
text(0, dim(morph.dor.1)[1]-10, "A", cex = 2)

# 2) Without Allometry: PC1 lateral view
par(mar = c(0, 0, 1, 0))
plot(c(0, dim(morph.lat.1)[2]), c(0, dim(morph.lat.1)[1]), type = "n", axes = F, xlab = "", ylab = "", asp = T) 
rasterImage(morph.lat.1, 0, 0, dim(morph.lat.1)[2], dim(morph.lat.1)[1])
text(0, dim(morph.lat.1)[1]-10, "B", cex = 2)

# 3) With Allometry: PC1 dorsal view
par(mar = c(0, 0, 1, 0)) 
plot(c(0, dim(res.dor.1)[2]), c(0, dim(res.dor.1)[1]), type = "n", axes = F, xlab = "", ylab = "", asp = T) 
rasterImage(res.dor.1, 0, 0, dim(res.dor.1)[2], dim(res.dor.1)[1])
text(0, dim(res.dor.1)[1]-10, "C", cex = 2)

# 4) Without Allometry: PC1 dorsal view
par(mar = c(0, 0, 1, 0))  
plot(c(0, dim(res.lat.1)[2]), c(0, dim(res.lat.1)[1]), type = "n", axes = F, xlab = "", ylab = "", asp = T) 
rasterImage(res.lat.1, 0, 0, dim(res.lat.1)[2], dim(res.lat.1)[1])
text(0, dim(res.lat.1)[1]-10, "D", cex = 2)

```

### Make figure comparing PC2 of with-allometry vs residuals
NOTE: would be best if residuals were generated in script 03
